started refactoring, no longer works
