+++ webApp/server/usersController.js
@@ -1,0 +1,5 @@
+var User = require('./usersModel');
+
+module.exports = {
+
+};
