+++ webApp/server/usersModel.js
@@ -1,0 +1,7 @@
+var mongoose = require('mongoose');
+
+var userSchema = mongoose.Schema({
+  name: String
+});
+
+module.exports = mongoose.model('User', userSchema);
